api/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-DotEnv==0.1.2
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.3
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
multidict==4.7.6
mysql-connector-python==8.0.20
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.10.0
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.6.0
